moved auth functions to auth.js
